Correct Sensor Naming
Hardware variables were mislabeled. I2C addresses for the ADCs that
read the left and right thermocouples were switched. As the hardware is
already in the field, addresses are corrected when the data is uploaded
to the SQL db.

Previously it was thought that the left and right temperature, relative
humidity and fans were incorrect… Mistakes were made, but have been
corrected.